set cmds and envs based on artifact storage
